Version 0.15 - Updates
Updating code. Main goal is to have the script either create and html
file as output, or output a cStringIO object that can then be .write'd
to some other file, so that the PythonParse class has more utility than
just a CLI tool. Might even get it to be used on a webserver, who knows.